test: Migrate to type inference test
