--- tests/Stubs/Symfony/Component/Form/DataClass.php
@@ -1,14 +1,0 @@
-<?php declare(strict_types = 1);
-
-namespace PHPStan\Stubs\Symfony\Component\Form;
-
-class DataClass
-{
-
-	/** @var int */
-	public $foo;
-
-	/** @var string */
-	public $bar;
-
-}
