--- tests/Stubs/Symfony/Component/Form/DataClassType.php
@@ -1,34 +1,0 @@
-<?php declare(strict_types = 1);
-
-namespace PHPStan\Stubs\Symfony\Component\Form;
-
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\NumberType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-/**
- * @extends AbstractType<DataClass>
- */
-class DataClassType extends AbstractType
-{
-
-	public function buildForm(FormBuilderInterface $builder, array $options): void
-	{
-		$builder
-			->add('foo', NumberType::class)
-			->add('bar', TextType::class)
-		;
-	}
-
-	public function configureOptions(OptionsResolver $resolver): void
-	{
-		$resolver
-			->setDefaults([
-				'data_class' => DataClass::class,
-			])
-		;
-	}
-
-}
